NVCC

+++ test/hipcub/test_hipcub_device_adjacent_difference.cpp
@@ -377,7 +377,7 @@ private:
     class check_output
     {
     public:
-        __device__ check_output(flag_type* incorrect_flag, size_t current_index, size_t* counter)
+        __device__ check_output(flag_type* incorrect_flag, size_t current_index, unsigned long long int* counter)
             : current_index_(current_index), incorrect_flag_(incorrect_flag), counter_(counter)
         {}
 
@@ -399,7 +399,7 @@ private:
     private:
         size_t         current_index_;
         flag_type*     incorrect_flag_;
-        unsigned long* counter_;
+        unsigned long long int* counter_;
     };
 
 public:
@@ -409,7 +409,7 @@ public:
     using iterator_category = std::random_access_iterator_tag;
     using difference_type = std::ptrdiff_t;
 
-    __host__ __device__ check_output_iterator(flag_type* const incorrect_flag, size_t* const counter)
+    __host__ __device__ check_output_iterator(flag_type* const incorrect_flag, unsigned long long int* const counter)
         : current_index_(0), incorrect_flag_(incorrect_flag), counter_(counter)
     {}
 
@@ -490,7 +490,7 @@ struct flag_expected_op : public std::binary_function<T, T, discard_write<T>>
 private:
     size_t     current_index_;
     flag_type* incorrect_flag_;
-    size_t*    counter_;
+    unsigned long long int*    counter_;
 };
 
 using HipcubDeviceAdjacentDifferenceLargeTestsParams
@@ -606,8 +606,8 @@ TYPED_TEST(HipcubDeviceAdjacentDifferenceLargeTests, LargeIndicesAndOpOnce)
         {
             SCOPED_TRACE(testing::Message() << "with size = " << size);
 
-            flag_type* d_incorrect_flag;
-            size_t*    d_counter;
+            flag_type*              d_incorrect_flag;
+            unsigned long long int* d_counter;
             HIP_CHECK(
                 test_common_utils::hipMallocHelper(&d_incorrect_flag, sizeof(*d_incorrect_flag)));
             HIP_CHECK(test_common_utils::hipMallocHelper(&d_counter, sizeof(*d_counter)));
@@ -661,8 +661,8 @@ TYPED_TEST(HipcubDeviceAdjacentDifferenceLargeTests, LargeIndicesAndOpOnce)
                                                    stream));
 
             // Copy output to host
-            flag_type incorrect_flag;
-            size_t    counter;
+            flag_type              incorrect_flag;
+            unsigned long long int counter;
             HIP_CHECK(hipMemcpy(&incorrect_flag,
                                 d_incorrect_flag,
                                 sizeof(incorrect_flag),
